requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Vector DB & Embeddings
chromadb>=1.3.0
langchain==0.1.0
langchain-community==0.0.10

# HTTP & API
requests==2.31.0
httpx>=0.27.0

# Testing & Automation
selenium==4.16.0
pytest==7.4.3
pytest-xdist==3.5.0
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tenacity==8.2.3

# Document Processing
pypdf==3.17.4
markdown==3.5.1
beautifulsoup4==4.12.2

# Logging & Monitoring
structlog==24.1.0

# UI Framework
streamlit==1.29.0
